This Code takes inputs and converts them into binary numbers.
This Code takes inputs and converts them into binary numbers. They are
in 16 bit form.

PLEASE DONT DOWNLOAD THIS CODE FOR A CSE CLASS AND USE IT AS YOURS...

DOWNLOADING THIS CODE AND USING IT AS YOURS WILL NOT HELP ANY
STUDENT...ESPECIALLY NOT CSE MAJORS

THIS CODE IS FOR HELP AND I KINDA DIDN'T WANT IT ON MY COMPUTER SO I
THREW IT UP HERE LOL